requirements.txt
addict==2.4.0
aiofiles==0.8.0
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
albumentations==1.1.0
alembic==1.8.0
anyio==3.6.1
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
asyncssh==2.11.0
atpublic==3.0.1
attrs==21.4.0
backcall==0.2.0
backports.cached-property==1.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
boto3==1.24.2
botocore==1.27.2
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028895253/work
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
chumpy==0.70
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
coverage==6.4.1
cryptography==37.0.2
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==0.29.30
dacite==1.6.0
databricks-cli==0.16.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dictdiffer==0.9.0
dill==0.3.5.1
diskcache==5.4.0
distro==1.7.0
dnspython==2.2.1
docker==5.0.3
dpath==2.0.6
dulwich==0.20.42
dvc==2.12.1
dvc-data==0.0.19
dvc-objects==0.0.19
dvc-render==0.0.6
dvclive==0.8.2
entrypoints==0.4
eventlet==0.33.1
executing==0.8.3
fastjsonschema==2.15.3
fiftyone==0.16.2
fiftyone-brain==0.8.2
fiftyone-db==0.3.0
flake8==4.0.1
Flask==2.1.2
flatbuffers==2.0
flatten-dict==0.4.2
flufl.lock==7.0
fonttools==4.33.3
freetype-py==2.3.0
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
glob2==0.7
grandalf==0.6
graphql-core==3.1.7
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
interrogate==1.5.0
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==4.3.21
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
json-tricks==3.15.5
jsonschema==4.6.0
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kaleido==0.2.1
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
mailchecker==4.1.17
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mlflow==1.26.1
mmcv-full==1.5.2
mmdet==2.25.0
-e git+https://github.com/open-mmlab/mmpose.git@84e949b5d7740e8be0692ac1f0aef9a99048a4fe#egg=mmpose&subdirectory=../../../source/mmpose
mongoengine==0.20.0
motmetrics==1.2.5
motor==2.5.1
multidict==6.0.2
munkres==1.1.4
nanotime==0.5.2
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
ndg-httpsclient==0.5.1
ndjson==0.3.1
nest-asyncio==1.5.5
networkx==2.8.2
notebook==6.4.12
numexpr @ file:///tmp/build/80754af9/numexpr_1640704208950/work
numpy==1.22.4
oauthlib==3.2.0
onnx==1.11.0
onnxruntime==1.11.1
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patool==1.12
pexpect==4.8.0
phonenumbers==8.12.49
pickleshare==0.7.5
Pillow==9.1.1
plotly==4.14.3
pluggy==1.0.0
poseval @ git+https://github.com/svenkreiss/poseval.git@3128c5cbcf90946e5164ff438ad651e113e64613
pprintpp==0.4.0
priority==2.0.0
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.2
prompt-toolkit==3.0.29
protobuf==4.21.1
psutil==5.9.1
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pycocotools==2.0.4
pycodestyle==2.8.0
pycparser==2.21
pydot==1.4.2
pyflakes==2.4.0
pygit2==1.9.2
pyglet==1.5.26
Pygments==2.12.0
pygtrie==2.4.2
PyJWT==2.4.0
pymongo==3.12.3
PyOpenGL==3.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrender==0.1.45
pyrsistent==0.18.1
pytest==7.1.2
pytest-runner==6.0.0
python-benedict==0.25.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==0.20.0
python-fsutil==0.6.1
python-multipart==0.0.5
python-slugify==6.1.2
pytorch-ignite==0.5.0.dev20220606
pytube==12.1.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.0.0
qudida==0.0.4
querystring-parser==1.2.4
requests==2.27.1
retrying==1.3.3
rfc3986==1.5.0
rich==12.4.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
scmrepo==0.0.25
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
Send2Trash==1.8.0
sentinel==0.3.0
Shapely==1.8.2
shortuuid==1.0.9
shtab==1.5.4
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
smplx==0.1.28
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
sqlparse==0.4.2
sse-starlette==0.10.3
sseclient-py==1.7.2
stack-data==0.2.0
starlette==0.16.0
strawberry-graphql==0.96.0
tabulate==0.8.9
terminado==0.15.0
terminaltables==3.1.10
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
torch==1.10.0+cu111
torchsummary==1.5.1
torchvision==0.11.0+cu111
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm==4.64.0
traitlets==5.2.2.post1
trimesh==3.12.5
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
universal-analytics-python3==1.1.1
urllib3==1.26.9
voluptuous==0.13.1
voxel51-eta==0.7.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.14.1
wsproto==1.1.0
xdoctest==1.0.0
xmltodict==0.13.0
xtcocotools==1.12
yapf==0.32.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.8.0
